+++ geowatch/mlops/manager.py
@@ -1135,8 +1135,8 @@ class UserAbort(Exception):
     pass
 
 
-__config__ = ManagerConfig
-__config__.main = main
+__cli__ = ManagerConfig
+__cli__.main = main
 
 
 if __name__ == '__main__':